code cleanup and fix

+++ libusbstm32/usbdev/Class/cdc/inc/usbd_cdc_core.h
@@ -116,39 +116,26 @@
 #define SEND_BREAK                              0x23
 #define NO_CMD                                  0xFF
 
-/**
-  * @}
-  */ 
 
+/*---------------------------------------------------------------------*/
 
 /** @defgroup USBD_CORE_Exported_TypesDefinitions
   * @{
   */
 typedef struct _CDC_IF_PROP
 {
-  uint16_t (*pIf_Init)     (void);   
-  uint16_t (*pIf_DeInit)   (void);   
-  uint16_t (*pIf_Ctrl)     (uint32_t Cmd, uint8_t* Buf, uint32_t Len);
-  uint16_t (*pIf_DataTx)   (const uint8_t* Buf, uint32_t Len);
-  uint16_t (*pIf_DataRx)   (uint8_t* Buf, uint32_t Len);
+  int (*pIf_Init)     (void);
+  int (*pIf_DeInit)   (void);
+  int (*pIf_Ctrl)     (uint32_t Cmd, uint8_t* Buf, uint32_t Len);
+  int (*pIf_DataTx)   (const uint8_t* Buf, uint32_t Len);
+  int (*pIf_DataRx)   (uint8_t* Buf, uint32_t Len);
 }
 CDC_IF_Prop_TypeDef;
-/**
-  * @}
-  */ 
 
+/*---------------------------------------------------------------------*/
 
 const USBD_Class_cb_TypeDef* cdc_class_init( const CDC_IF_Prop_TypeDef  *app_fops );
 
-
+/*---------------------------------------------------------------------*/
 
 #endif  // __USB_CDC_CORE_H_
-/**
-  * @}
-  */ 
-
-/**
-  * @}
-  */ 
-  
-/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
